requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
ultralytics==8.2.0
opencv-python==4.8.1.78
easyocr==1.7.0
numpy==1.26.4
pydantic~=2.5
torch==2.0.1 
torchvision==0.15.2
